Adding LineMapper.java file
This uploaded file contains the foundation for the contents of the entire project. The file currently contains hardcoded sample file contents represented as 2D string arrays (a good approach for easily accessing and manipulating file contents. The file also contains many print statements to assist in tracing the behaviour of the program. Please only copy and upload to your own branches if you wish to continue working on this code.